Move functions out into own 'chunk' module
